refactor: extract billing queryFn into queryOptions factories
- Add subscriptionStatus factory to billingQueries in billing-keys.ts
- Add subscriptionQueries (list, detail) to billing-keys.ts
- Simplify use-billing.ts hooks to delegate to factories
- Remove now-unused createClient/getCachedUser/logger imports from use-billing.ts
- Re-export subscriptionQueries for downstream consumers
