+++ src/hooks/api/query-keys/billing-keys.ts
@@ -20,6 +20,7 @@ import type {
 	CreateRentSubscriptionRequest,
 	FailedPaymentAttempt,
 	RentSubscriptionResponse,
+	SubscriptionStatusResponse,
 	UpdateSubscriptionRequest
 } from '#types/api-contracts'
 
@@ -250,6 +251,152 @@ export const billingQueries = {
 			},
 			enabled: !!subscriptionId,
 			staleTime: 30 * 1000
+		}),
+
+	subscriptionStatus: (options?: { enabled?: boolean }) =>
+		queryOptions({
+			queryKey: billingKeys.subscriptionStatus(),
+			queryFn: async (): Promise<SubscriptionStatusResponse> => {
+				const supabase = createClient()
+				const user = await getCachedUser()
+				if (!user) {
+					logger.warn('Subscription status check: no user')
+					throw new Error('Not authenticated')
+				}
+
+				// Get stripe_customer_id from users table
+				const { data: userData, error: userError } = await supabase
+					.from('users')
+					.select('stripe_customer_id')
+					.eq('id', user.id)
+					.single()
+
+				if (userError) handlePostgrestError(userError, 'users')
+
+				const stripeCustomerId = userData?.stripe_customer_id ?? null
+
+				if (!stripeCustomerId) {
+					logger.debug('No stripe_customer_id, returning null status')
+					return {
+						subscriptionStatus: null,
+						stripeCustomerId: null,
+						stripePriceId: null,
+						currentPeriodEnd: null,
+						cancelAtPeriodEnd: false
+					} satisfies SubscriptionStatusResponse
+				}
+
+				// Query stripe.subscriptions for real subscription status
+				const { data: subData, error: subError } = await supabase
+					.rpc('get_subscription_status', { p_customer_id: stripeCustomerId })
+
+				if (subError) {
+					logger.debug('get_subscription_status RPC not available, falling back to leases', {
+						error: subError.message
+					})
+
+					const { data: leaseData } = await supabase
+						.from('leases')
+						.select('stripe_subscription_status')
+						.eq('owner_user_id', user.id)
+						.not('stripe_subscription_id', 'is', null)
+						.order('created_at', { ascending: false })
+						.limit(1)
+						.maybeSingle()
+
+					const leaseStatus = leaseData?.stripe_subscription_status as string | null
+
+					return {
+						subscriptionStatus: (leaseStatus ?? null) as SubscriptionStatusResponse['subscriptionStatus'],
+						stripeCustomerId,
+						stripePriceId: null,
+						currentPeriodEnd: null,
+						cancelAtPeriodEnd: false
+					} satisfies SubscriptionStatusResponse
+				}
+
+				const sub = (Array.isArray(subData) ? subData[0] : subData) as Record<string, unknown> | null
+
+				const status = (sub?.status as string) ?? null
+				logger.debug('Subscription status from stripe.subscriptions', { status })
+
+				return {
+					subscriptionStatus: status as SubscriptionStatusResponse['subscriptionStatus'],
+					stripeCustomerId,
+					stripePriceId: (sub?.price_id as string) ?? null,
+					currentPeriodEnd: (sub?.current_period_end as string) ?? null,
+					cancelAtPeriodEnd: (sub?.cancel_at_period_end as boolean) ?? false
+				} satisfies SubscriptionStatusResponse
+			},
+			staleTime: 5 * 60 * 1000,
+			gcTime: 10 * 60 * 1000,
+			enabled: options?.enabled ?? true
+		})
+}
+
+/**
+ * Subscription query options for rent subscriptions (lease-based)
+ */
+export const subscriptionQueries = {
+	list: () =>
+		queryOptions({
+			queryKey: subscriptionsKeys.list(),
+			queryFn: async (): Promise<RentSubscriptionResponse[]> => {
+				const supabase = createClient()
+				const { data, error } = await supabase
+					.from('leases')
+					.select('id, stripe_subscription_id, stripe_subscription_status, rent_amount, rent_currency, primary_tenant_id, unit_id')
+					.not('stripe_subscription_id', 'is', null)
+					.order('created_at', { ascending: false })
+				if (error) handlePostgrestError(error, 'leases')
+				return (data ?? []).map((row): RentSubscriptionResponse => ({
+					id: row.id,
+					leaseId: row.id,
+					tenantId: row.primary_tenant_id ?? '',
+					ownerId: '',
+					stripeSubscriptionId: row.stripe_subscription_id ?? '',
+					stripeCustomerId: '',
+					amount: row.rent_amount ?? undefined,
+					currency: row.rent_currency ?? 'USD',
+					billingDayOfMonth: 1,
+					status: row.stripe_subscription_status ?? 'unknown',
+					platformFeePercentage: 0,
+					createdAt: '',
+					updatedAt: ''
+				}))
+			},
+			staleTime: 30 * 1000
+		}),
+
+	detail: (id: string) =>
+		queryOptions({
+			queryKey: subscriptionsKeys.detail(id),
+			queryFn: async (): Promise<RentSubscriptionResponse> => {
+				const supabase = createClient()
+				const { data, error } = await supabase
+					.from('leases')
+					.select('id, stripe_subscription_id, stripe_subscription_status, rent_amount, rent_currency, primary_tenant_id, unit_id')
+					.eq('id', id)
+					.single()
+				if (error) handlePostgrestError(error, 'leases')
+				const row = data!
+				return {
+					id: row.id,
+					leaseId: row.id,
+					tenantId: row.primary_tenant_id ?? '',
+					ownerId: '',
+					stripeSubscriptionId: row.stripe_subscription_id ?? '',
+					stripeCustomerId: '',
+					amount: row.rent_amount ?? undefined,
+					currency: row.rent_currency ?? 'USD',
+					billingDayOfMonth: 1,
+					status: row.stripe_subscription_status ?? 'unknown',
+					platformFeePercentage: 0,
+					createdAt: '',
+					updatedAt: ''
+				} satisfies RentSubscriptionResponse
+			},
+			enabled: !!id
 		})
 }
